+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.2.0] - in development
+## [0.2.0] - 2024-10-27
 
 - MSRV bumped to 1.75.0 (because of `-> impl Trait` in a trait method).
 - Renamed `AsMutDeserializer` -> `AsTransientDeserializer`.